+++ sdk/src/protocols/aave/v2/tests/mainnet/borrow.test.ts
@@ -25,7 +25,7 @@ describe.skip("aaveV2", () => {
         Chain.eth,
         contracts.mainnet.usdc,
         parseUnits("10000", 6),
-        contracts.mainnet.balancer.vault
+        contracts.mainnet.balancerV2.vault
       )
       await expect(
         kit.asMember.usdc.approve(
@@ -184,7 +184,7 @@ describe.skip("aaveV2", () => {
         Chain.eth,
         contracts.mainnet.usdc,
         parseUnits("10000", 6),
-        contracts.mainnet.balancer.vault
+        contracts.mainnet.balancerV2.vault
       )
 
       await expect(